change on hotel list and detail api

+++ app/src/main/java/com/example/final_project/models/hotel/HotelDetail.kt
@@ -34,10 +34,10 @@ data class HotelDetail(
     val maxPrice: Long,
 
     @SerializedName("latitude")
-    val latitude: Long,
+    val latitude: Double,
 
     @SerializedName("longitude")
-    val longitude: Long,
+    val longitude: Double,
 
     @SerializedName("cluster")
     val cluster: Long,
